Moved panels as a subfolder to components
