+++ assets/src/edit-story/components/inspector/design/designInspector.js
@@ -22,7 +22,7 @@ import styled from 'styled-components';
 /**
  * Internal dependencies
  */
-import { LayerPanel, ColorPresetPanel } from '../../../panels';
+import { LayerPanel, ColorPresetPanel } from '../../panels';
 import DesignPanels from './designPanels';
 
 const Wrapper = styled.div`
