+++ assets/src/edit-story/components/inspector/prepublish/prepublishInspector.js
@@ -22,7 +22,7 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { SimplePanel } from '../../../panels/panel';
+import { SimplePanel } from '../../panels/panel';
 
 function PrepublishInspector() {
 	return (
