+++ assets/src/edit-story/components/panels/font.js
@@ -28,9 +28,9 @@ import { __ } from '@wordpress/i18n';
 /**
  * Internal dependencies
  */
-import { InputGroup, SelectMenu } from '../components/form';
-import { useFont } from '../app';
-import { MIN_FONT_SIZE, MAX_FONT_SIZE } from '../constants';
+import { InputGroup, SelectMenu } from '../form';
+import { useFont } from '../../app';
+import { MIN_FONT_SIZE, MAX_FONT_SIZE } from '../../constants';
 import { SimplePanel } from './panel';
 import getCommonValue from './utils/getCommonValue';
