+++ assets/src/edit-story/components/panels/panel/handle.js
