+++ assets/src/edit-story/components/panels/panel/simplePanel.js
