+++ assets/src/edit-story/components/panels/utils/getCommonValue.js
